+++ src/components/DailyIncreaseChart/DailyIncrease.js
